+++ index.html
@@ -228,7 +228,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-<script src="js/functions.js"></script>
 <script type=module src="js/main.js"></script>
 </body>
 </html>
